Fix bug in roster fit algorithm
While swapping players different positions to make room for a new
player, we had selected the same player twice.  This messed up the
algorithm because the player had already moved, so there wasn't
a roster spot for a given position.  Keep track of the players
we moved when fitting so that we don't move the same player twice.